requirements.txt
Flask==2.3.2
ffmpeg_python==0.2.0
protobuf==3.20.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
stable_ts==2.17.5
torch==2.5.1
google-cloud-storage==1.44.0
google-cloud-aiplatform==1.71.0